:recycle: Refactor python packages to work without bazel
Previous CLI
============
bazel-bin/src/exporters/reljam --root demo/petshop model /petshop PetShopModel
bazel-bin/src/sysl/sysl --root demo/petshop textpb -o out/petshop.txt /petshop

New CLI
=======
> cd <repo_root>
> export PYTHONPATH=`pwd`/src
> python -m sysl  --root demo/petshop textpb -o out/petshop.txt /petshop
> python -m sysl.reljam  --root demo/petshop model /petshop PetShopModel

Next up:
=======
setup.py to make distribution and installable scripts
