+++ src/sysl/core/syslints.py
@@ -5,8 +5,8 @@
 import itertools
 import re
 
-from sysl.sysl import syslloader
-from sysl.sysl import syslx
+from sysl.core import syslloader
+from sysl.core import syslx
 
 from sysl.util import diagutil
 from sysl.util import writer
