+++ src/sysl/core/syslparse.py
@@ -18,7 +18,7 @@ import re
 
 from sysl.proto import sysl_pb2
 
-from sysl.sysl import syslx
+from sysl.core import syslx
 
 from sysl.util import simple_parser
 from sysl.util import rex
