+++ src/sysl/exporters/java/model.py
@@ -10,7 +10,7 @@ import warnings
 
 from sysl.proto import sysl_pb2
 
-from sysl.sysl import syslx
+from sysl.core import syslx
 
 from sysl.util import datamodel
 from sysl.util import java
