+++ src/sysl/exporters/json_out/serializer.py
@@ -1,6 +1,6 @@
 from sysl.proto import sysl_pb2
 
-from sysl.sysl import syslx
+from sysl.core import syslx
 
 from sysl.util import datamodel
 from sysl.util import java
