+++ src/sysl/exporters/swagger/swagger.py
@@ -1,7 +1,7 @@
 # -*- encoding: utf-8 -*-
 import yaml
 
-from sysl.sysl import syslx
+from sysl.core import syslx
 
 from sysl.util import datamodel
 from sysl.util import java
