+++ src/sysl/exporters/xml/xsd.py
@@ -2,7 +2,7 @@ import contextlib
 
 from sysl.proto import sysl_pb2
 
-from sysl.sysl import syslx
+from sysl.core import syslx
 
 from sysl.util import datamodel
 from sysl.util import java
