+++ src/sysl/util/java.py
@@ -11,7 +11,7 @@ import uuid
 
 from sysl.proto import sysl_pb2
 
-from sysl.sysl import syslx
+from sysl.core import syslx
 
 import datamodel
 import scopes
