+++ src/sysl/util/scopes.py
@@ -1,6 +1,6 @@
 from sysl.proto import sysl_pb2
 
-from sysl.sysl import syslx
+from sysl.core import syslx
 
 
 class Scope(object):
